# data-agent/pg-agent/swagger.yaml
openapi: 3.0.3
info:
  title: PG-Agent NLQ API
  description: "Natural Language Query endpoint for PG-Agent. Use this spec to integrate with the /api/nlq/ask endpoint."
  version: 1.0.0
servers:
  - url: http://10.10.8.206:8001
paths:
  /ask:
    post:
      summary: Chatbot-friendly Q&A endpoint
      operationId: askSimplified
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NLQRequest'
            example:
              question: "How many claims are there for 2025?"
              limit: 100
      responses:
        '200':
          description: Simplified response with question, answer, and generated SQL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimplifiedAskResponse'
        '400':
          description: Validation or generation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/nlq/ask:
    post:
      summary: Natural Language Query (full technical response)
      operationId: askNLQFull
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NLQRequest'
      responses:
        '200':
          description: Full response with SQL, columns, row data, and token usage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NLQResponse'
        '400':
          description: Validation or generation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    SimplifiedAskResponse:
      type: object
      properties:
        question:
          type: string
          description: The user's question.
        answer:
          type: string
          description: AI-generated natural language answer (chatbot-friendly).
        generated_sql:
          type: string
          description: SQL generated and executed to answer the question.
        session_id:
          type: string
          description: Per-chat identifier used to store/reload conversation history.
        messages:
          type: array
          items:
            $ref: '#/components/schemas/NLQChatMessage'
    NLQRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          description: The natural language question to ask about the database.
          example: "How many open claims are there for 2024?"
        limit:
          type: integer
          description: Maximum number of rows to return (default 100).
          default: 100
        session_id:
          type: string
          description: If provided, the server will load stored chat history for this session (process-local).
        messages:
          type: array
          items:
            $ref: '#/components/schemas/NLQChatMessage'
    NLQResponse:
      type: object
      properties:
        question:
          type: string
        generated_sql:
          type: string
          description: SQL generated by the LLM to answer the question.
        row_count:
          type: integer
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            description: Result rows as JSON objects (column name -> value).
        model:
          type: string
          description: LLM model used for the generation.
        prompt_tokens:
          type: integer
        reply_tokens:
          type: integer
    NLQChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum:
            - user
            - assistant
        content:
          type: string
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message.